examples/the-mastermind/pyproject.toml
[project]
name = "the-mastermind"
version = "0.1.0"
description = "The Mastermind - HTTP Agent for Crosswind Heist Crew"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
